Renames the util-function files to follow naming convention
