+++ src/util/findRouteByComponent.js
